refactor: generateSwiftCode → swiftCodeGen
